% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_plot}
\alias{create_plot}
\title{Create Histogram of a named variable}
\usage{
create_plot(data, variable)
}
\arguments{
\item{data}{tibble to be utilized}

\item{variable}{name of the variable to be plotted}
}
\value{
ggplot object
}
\description{
Create Histogram of a named variable
}
