% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scenario_1.R
\name{evaluate_process}
\alias{evaluate_process}
\title{Evaluate the made analysis}
\usage{
evaluate_process(analysis)
}
\arguments{
\item{analysis}{made before}
}
\value{

}
\description{
Evaluate the made analysis
}
