% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scenario_1.R
\name{export_result}
\alias{export_result}
\title{Export the Result artifact}
\usage{
export_result(evaluated_process)
}
\arguments{
\item{if}{process was evaluated succesful}
}
\value{
TRUE as stub
}
\description{
Export the Result artifact
}
