% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{icu_example_dataset}
\alias{icu_example_dataset}
\title{Neuro-ICU Synthetic Dataset}
\description{
This dataset is synthetically generated and gives example of real values
}
\concept{example_datasets}
\section{Output Columns:}{
\describe{
\item{gupi}{Numeric.
\code{gupi} or "General Unified Persistent Identifier" identifies a subject across different datasets.
\itemize{}
}
\item{timestamp}{dttm
Timestamp of recorded value(s)
\itemize{}
}
\item{ICP}{Numeric.
Intracranial Pressure measured in mmHg
\itemize{}
}
}
}
