% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scenario_1.R
\name{import_data}
\alias{import_data}
\title{Imports data linked with the question}
\usage{
import_data(question, file)
}
\arguments{
\item{question}{question string, which is linked in the system}

\item{file}{path to raw triplet file}
}
\value{
imported data
}
\description{
Imports data linked with the question
}
