% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scenario_1.R
\name{make_analysis}
\alias{make_analysis}
\title{The analysis makes a report, in the example case a visualization of cpp}
\usage{
make_analysis(data, notebook_file)
}
\arguments{
\item{data}{returned from import data}

\item{notebook_file}{the report file which is knitted}
}
\value{

}
\description{
The analysis makes a report, in the example case a visualization of cpp
}
