% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{my_function}
\alias{my_function}
\title{Increment Template Function}
\usage{
my_function(input)
}
\arguments{
\item{input}{integer to increment}
}
\value{
output numeric - increased by 1
}
\description{
Increment Template Function
}
\examples{
my_function(1)

}
