% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_scenario_2.R
\name{sample_data}
\alias{sample_data}
\title{Samples data linked with the question from scenario 1}
\usage{
sample_data(data)
}
\arguments{
\item{data}{the tibble which is subsampled}
}
\value{
sampled data, which should be reprodcicble
}
\description{
Samples data linked with the question from scenario 1
}
