% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\docType{data}
\name{scenario_1}
\alias{scenario_1}
\title{Plan of scenario_1}
\format{
An object of class \code{drake_plan} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
scenario_1
}
\description{
Plan of scenario_1
}
\keyword{datasets}
