% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{scenario_wrapper}
\alias{scenario_wrapper}
\title{Plan of scenarios,
This is where you set up your workflow plan,
a data frame with the steps of your analysis.}
\usage{
scenario_wrapper()
}
\description{
Plan of scenarios,
This is where you set up your workflow plan,
a data frame with the steps of your analysis.
}
