% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_plot}
\alias{create_plot}
\title{Create Histogram with Mapping Ozone and Binwidth = 10}
\usage{
create_plot(data)
}
\arguments{
\item{data}{}
}
\value{
ggplot object
}
\description{
Create Histogram with Mapping Ozone and Binwidth = 10
}
